feat: Abstract collections to make path information routable
This is a first stab at a common pathing problem in flows: POSIX
paths map to different locations than Globus Shared Endpoint paths.

Mainly, this introduces the SharedCollection path which tracks the POSIX
path associated with a Globus share. Paths can then be 'translated' to
either posix or globus paths as needed.