chore: Remove code no longer used in processing
For way too long, we kept around the old versions of things instead of
getting rid of them. These are far out of date and no longer useful at
this point in time. Even reprocessing is far enough out of date, it doesn't
serve a useful purpose.

Init shortcuts also removed, as those weren't adding a whole lot and
obfuscating the locations of modules.