Restructure how paths are handled in boost corr flow
Previously, every file was transferred for the flow. The main problem
was that the hdf metadata file was not being transferred into the
input directory with the other "raw" files required for processing.
The misconception was because the 'hdf' metadata file is not explicitly
passed to corr, but rather just needs to be in the same directory,
and this was not the case (And corr wasn't reporting an error).

New structure passes the whole input directory, both the raw files
and the metadata hdf together. The metadata hdf is no longer needed,
only the 'raw' file to pass to corr. Handling of the qmap file is
unchanged.

Some internal refactor was done to simplify how directories are handled
instead of individual files, which meant we could remove the rigaku
functions. Additionally, one tool was added to track corr metadata.